moved XmlExport, some formatting
