+++ pom.xml
@@ -47,10 +47,11 @@
             <artifactId>json</artifactId>
             <version>20160810</version>
         </dependency>
-      <dependency>
-        <groupId>org.junit.jupiter</groupId>
-        <artifactId>junit-jupiter-api</artifactId>
-        <version>RELEASE</version>
-      </dependency>
+        <dependency>
+          <groupId>org.junit.jupiter</groupId>
+          <artifactId>junit-jupiter-api</artifactId>
+          <version>RELEASE</version>
+          <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
