+++ src/main/java/scadinspect/data/scaddoc/XmlExport.java
@@ -1,14 +1,15 @@
-package scadinspect.data.scaddoc.properties;
+package scadinspect.data.scaddoc;
 
 /**
  * Created by eric on 24.03.17.
  */
 
-import java.io.File;
+import java.io.StringWriter;
 import java.util.Collection;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.OutputKeys;
 import javax.xml.transform.Transformer;
 import javax.xml.transform.TransformerException;
 import javax.xml.transform.TransformerFactory;
@@ -16,11 +17,12 @@ import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
-import scadinspect.data.scaddoc.Module;
+import scadinspect.data.scaddoc.properties.PairProperty;
+import scadinspect.data.scaddoc.properties.Property;
 
 public class XmlExport {
 
-  public void getXml(Collection<Module> modules)
+  public String getXml(Collection<Module> modules)
       throws ParserConfigurationException, TransformerException {
 
     try {
@@ -34,26 +36,29 @@ public class XmlExport {
       doc.appendChild(rootElement);
 
       for (Module module : modules) {
-        //TODO Stuff
         rootElement.appendChild(createModuleNode(module, doc));
       }
 
       // write the content into xml file
       TransformerFactory transformerFactory = TransformerFactory.newInstance();
       Transformer transformer = transformerFactory.newTransformer();
+      transformer.setOutputProperty(OutputKeys.METHOD, "xml");
+      transformer.setOutputProperty(OutputKeys.INDENT, "yes");
+      transformer.setOutputProperty("{http://xml.apache.org/xslt}indent-amount", "2");
       DOMSource source = new DOMSource(doc);
-      StreamResult result = new StreamResult(new File("C:\\file.xml"));
+      StringWriter sw = new StringWriter();
+      StreamResult result = new StreamResult(sw);
 
       // Output to console for testing
       // StreamResult result = new StreamResult(System.out);
 
       transformer.transform(source, result);
 
-    } catch (ParserConfigurationException e) {
+      //return XML String + do some **magic**
+      return sw.getBuffer().toString().replaceAll("((?<=>)\\[)|(](?=<))","");
+
+    } catch (Exception e) {
       // TODO log parse failure
-      throw e;
-    } catch (TransformerException e) {
-      // TODO log transformer exception
       throw e;
     }
   }
@@ -66,7 +71,7 @@ public class XmlExport {
 
     for (Property property : properties) {
       Element key = document.createElement(property.getKey());
-      String value = "";
+      String value;
       if (property instanceof PairProperty) {
         PairProperty temp = (PairProperty) property;
         value = temp.getValue().getValue() + temp.getValue().getMetric();
